Fixed issue in the .babylon serializer (#16524)
When serializing to .babylon file we are setting the _isExpanded
property directly in the array object to flag if matrices indices are in
compact or expanded form. However, JSON.stringify will remove those
properties since it does not support properties in array objects. I'm
moving those flags to the serializationObject itself to allows it to be
properly exported. I'm also keeping the deserializer for it just in case
someone has a file that somehow has that property in the array.

Releated to Exporters PR:
https://github.com/BabylonJS/Exporters/pull/1137

---------

Co-authored-by: Raanan Weber <raananw+github@gmail.com>