Get WebAudio working on iOS when ringer switch is turned off (#16558)
WebAudio on iOS is disabled when the device's ringer switch is turned
off. This is confusing for users because non-WebAudio audio plays when
the ringer switch is off, which sets the expectation that WebAudio
should work with the ringer off as well.

This change fixes the issue by creating a silent HTMLAudioElement and
playing it when the audio engine receives a user gesture. This has the
side effect of enabling WebAudio on iOS, even when the ring switch is
off.