Add linear hex value to color picker when color is stored in linear mode (#16581)
Today if a material stores color in linear mode (Ex: PBR) the color
picker will do an auto conversion to gamma mode. This results in
confusion when copying the hex value into code (user needs to know to
convert it to linear when passing the hex)

This change adds the linear hex in the UX as well as a more informative
message instructing how to use the hex values.

Co-authored-by: Georgina <gehalper@microsoft.com>