bug fix - "moveWithCollisions" method parameter slideOnCollide==false now moves mesh to NEAR the collision point, instead of AT the collision point (#16587)
Issue - "moveWithCollisions" with the parameter slideOnCollide == false
used to move the mesh up to the point of collision. Due to floating
point inaccuracy, this sometimes meant just outside, and sometimes just
inside penetration with the collided mesh. When just inside, subsequent
calls to "moveWithCollisions" would snag on the mesh previously collided
with.

Solution - "moveWithCollisions" with the parameter slideOnCollide ==
false now moves the mesh NOT up to the point of collision, but up to a
point just a small distance short of that collision. The small distance
is based on "AbstractEngine.CollisionsEpsilon", the same distance
considered "close" by other aspects of the collision system.