Sandbox: fix pan speed when model was loaded via drag and drop (#16590)
There were previously 2 possible render loops depending on if you loaded
from a URL or drag & drop. The previous fix for camera pan speed only
applied to one of them, so if you didn't use a URL and just dragged and
dropped in, you wouldn't get the corrected behavior. This change
consolidates down to one render loop for both scenarios, so the camera
behavior only has to be defined in one place. Specifically, FilesInput
is no longer responsible for creating a render loop - it's now always
done by renderingZone.tsx in either case.