Viewer: Fix ignored options bug & camera framing with multiple models (#16604)
This PR primarily is to address this forum issue:
https://forum.babylonjs.com/t/babylon-viewer-v2/54317/82

The main problem is that a wrapper options object is constructed, but
does not correctly handle options that are just pass-throughs (e.g. the
ViewerElement doesn't do anything special with them, like map them to
attributes, which is the case for `useRightHandedSystem` for example),
and also does not correctly handle callback options (the callbacks
passed in are overwritten by the ViewerElement's own callbacks). To fix
this, I'm introducing an Options Proxy at the ViewerElement layer
similar to what was done in the ViewerFactory layer. While Proxy may be
slower for evaluating properties, this is not in the hot path (e.g.
options are not evaluated on a per-frame basis), so this should be a
fine usage of Proxy.

Once this change was in place, it exposed another issue, where the
ViewerFactory layer was writing to one of the options, which calls the
setter on the Proxy, which passes through to setting a property directly
on the target object. This was not intentional, it was just leftover
from how the options worked before the Proxy was introduced and got
missed. To fix this, instead of overwriting the `onInitialized` options
property in the ViewerFactory, it is handled in the options Proxy within
the ViewerFactory. To make sure this kind of error doesn't happen again,
I also made all the options readonly.

Finally this PR has one other unrelated but small fix, where in the case
of multiple models, when one is made "active," the camera is reset to
the bounds of that single model, not all loaded models.