Viewer: Add shadow support (#16459)
In the viewer, add the support for :
- shadow map,
- ibl shadows
- shadow quality selection `normal`, `high`

For now, ibl shadows are dissabled for meshes that have animations in
WebGPU even if the animation is paused as there is some issues.

The auto shadow selection will come in another PR.
The placement of the light will come in another PR.

Edit from @ryantrem: from local testing, these are the remaining issues
I'm seeing:
- [x] Shadow map shadows don't work when "statically" enabled (e.g.
`<babylon-viewer shadow-quality="normal">`).
- [ ] IBL shadows do not update when an animation is running. I assume
this is not supported in IBL shadows, but if that is the case, should we
disallow IBL shadows for animated models?
- [ ] IBL shadows seem to get into a bad state after I load a few
different models. This can be reproed in the Viewer test app by just
dragging and dropping a sequence of models with IBL shadows enabled.
After a few models, I start seeing wrong shadows or significant
rendering artifacts. I'm not sure if there is a missing accumulator
reset, or if this is a bug in IBL shadows or what.

---------

Co-authored-by: Alexandre Mottet <mottet@adobe.com>
Co-authored-by: Ryan Tremblay <ryan.tremblay@microsoft.com>