Let Mesh.bakeTransformIntoVertices use DeepImmutable<Matrix> (#16637)
I ran into a use case where I would like to transform a mesh by a
`DeepImmutable<Matrix>`, but the type checker complains because it does
not satisfy `Matrix`.

However, `Mesh.bakeTransformIntoVertices(transform)` does not need the
Matrix to be mutable.

This PR simply allows a `DeepImmutable<Matrix>` to be passed to
`Mesh.bakeTransformIntoVertices(transform)`