Dispose all entities of an AssetContainer (#16638)
There was a recent regression in `AssetContainer.dispose` where
basically only half the assets actually get disposed and removed from
the scene. I noticed this while testing some unrelated stuff in the
Viewer. This change includes:

- Fixing the regression in AssetContainer.dispose by re-introducing the
call to .slice(0) to copy the arrays.
- Adding a unit test for this AssetContainer scenario.
- Fixing the original Viewer bug I was looking at, where initial model
and env load are doubled (once by the Viewer respecting the current
options that includes the source and env, and once by an explicit call
that should have been removed when I fleshed out the config/options). 😬