instancedMesh: fix getWorldMatrix in case billboardMode of source mesh enabled (#16641)
The `_currentLOD` of InstancedMesh would be `_sourceMesh` if LOD of
source mesh not enabled, but if billboardMode of source mesh enabled,
`getWorldMatrix` would incorrectly use the billboardMode of the
`sourceMesh` to compute world matrix. This commit would add a check in
`getWorldMatrix` to skip this if `_currentLOD` is `_sourceMesh`.

Forum post:
<https://forum.babylonjs.com/t/billboardmode-of-sourcemesh-changes-instance-rotation/58469>
Playground: <https://playground.babylonjs.com/#1RQ8OR>