Improve error handling in CubeTexture (#16643)
See
https://forum.babylonjs.com/t/cubetexture-onerror-callback-not-called/58248?u=raananw

When loading a cube texture it is possible that an internal texture will
be created that will not be used by the cube texture and will not be
disposed. This happens, for example, when the file itself is not an env
texture. When that happens, it resides in the cache and returned if the
same URL is requested. Playground to reproduce -
https://playground.babylonjs.com/#633US3#10

This adds a local onError that disposes the failed internal texture,
thus taking it out of the cache