PBR material: Fix an incorrect #if evaluation in vertex shader (#16649)
`pbrBRDFFunctions` was not included in the PBR vertex shader, which
means that `#if BASE_DIFFUSE_MODEL != BRDF_DIFFUSE_MODEL_LAMBERT &&
BASE_DIFFUSE_MODEL != BRDF_DIFFUSE_MODEL_LEGACY` was not evaluated
correctly.

Note that `environmentBrdfSampler` is defined globally, but only in
fragment shader code. That's why I added a check on
**PBR_VERTEX_SHADER** to skip evaluating all `pbrBRDFFunctions`, except
for the definitions at the top.