Prevent microphone feedback loop (#16660)
Microphone sound sources are connected to the main audio bus by default,
and get output to the speakers. This causes an extremely loud feedback
loop that should be avoided.

This change adds an `outBusAutoDefault` option to the sound source,
defaulted to `true`, except for the microphone sound sources which
default to `false` so they don't automatically output to the speakers.