Viewer: add @babylonjs/materials as peerDependency (#16672)
With the introduction of shadows in the Viewer, we took a dependency on
`ShadowOnlyMaterial` from `@babylonjs/materials`, but missed adding it
as a `peerDependency`. This PR adds it, and addresses
https://forum.babylonjs.com/t/babylon-viewer-v2/54317/90.

However, while testing I noticed that what happens for existing projects
is that they may have `@babylonjs/core` <8.6.0, and then
`@babylonjs/materials` >=8.6.0 gets installed, and these two versions
are incompatible because of internal API breaking changes (casing on
function names). I expect hitting this issue would be common. We can
deal with it in the Viewer by making all @babylonjs/* dependencies be
^8.6.0, but it seems like the correct fix is to update
`@babylonjs/materials`'s dependency on `@babylonjs/core` to ^8.6.0.
@RaananW thoughts?

---------

Co-authored-by: Ryan Tremblay <ryan.tremblay@outlook.com>