Convert color objects to vectors when using json pointers (#16684)
glTF interactivity doesn't differentiate between color3 and vector3. As
the math nodes all work with vector3 and there are no specific Color
nodes (in interactivity) it is safe to convert any value received as
color to a vector object, and converting it back to color when saving it
in the babylon object.
The conversion from vector to color is already iplemented, the only
thing missing was the vcolor to vector when reading values.