RH camera fixes (#16691)
This fixes an issue we have with cameras when in right-handed system
where setting `rotation` or `rotationQuaterion` on a camera in
right-handed system actually sets the camera 180 flipped from what is
represented in the view matrix.

You can see this problem using this PG:
https://playground.babylonjs.com/#BDQ72M#1
On line 14, we are setting the `rotation` to `0,0,0`, but the rotation
from the world matrix of the camera is `0,PI,0`.

This change fixes the problem, but it is a breaking change for setting
the rotation properties since it will behave 180 flipped from before.
The default orientation of camera will still work as before (i.e.,
always point to +Z for both LH and RH) to maintain backwards
compatibility as much as possible.

The fix involves changing the reference point of camera. The code
mistakenly uses the same reference point of `0,0,1` for both LH and RH
view matrix. Babylon target camera's LH and RH view matrices are flipped
by Z. Thus, the reference point should also flip by Z for the math to
work.

There is a secondary problem in `setTarget` where it is hard-coded to
use LH math. This has been fixed to consider RH.