Add IReadonlyObservable (interface for receiver side of Observable) (#16702)
This PR adds an `IReadonlyObservable` interface which is the receiver
side of the Observable contract. I want to add this for two reasons:

1. The `Observable` class is type invariant (meaning it is not covariant
or contravariant). This means you cannot do something like `const
nodeObservable: Observable<Node> = new Observable<Mesh>()`, even though
a `Mesh` is a `Node`. This makes it hard to write code that should be
the same regardless of what type of `Node` you are trying to observe
(e.g. same code using `Scene.onNewMeshAddedObservable` and
`Scene.onNewTransformNodeAdded` for example). You can, but it requires
some ugly and type unsafe casting. This problem happens specifically
because `Observable` is not covariant. It's not covariant because it is
invariant because the type `T` is used as both inputs and outputs of the
contract, which means if you have `IObservable<T1>` and
`IObservable<T2>`, `T1` must exactly be `T2` (neither can be subtype of
the other). `IReadonlyObservable` is the receiver side of `Observable`
and only exposes `T` as outputs, meaning it is covariant, and therefore
`const nodeObservable: IReadonlyObservable<Node> = new
Observable<Mesh>()` is perfectly valid.
2. Separating out the receiver side of `Observable` makes it easier to
express intent in the API. For example, `Scene` should be the one firing
the `Scene.onNewMeshAddedObservable`, not anyone outside of the `Scene`.
However, since we current expose `onNewMeshAddedObservable` as an
`Observable<T>`, nothing stops code outside of `Scene` from calling
`notifyObservers`.

I used the name `IReadonlyObservable` because it seems like people will
likely understand what this means, and because if I recall @AmoebaChant
did something like this in another project. The downside of this name is
that it's harder to come up with a name for an interface for the "write"
side of `Observable` if we ever decide to do that (e.g. if we want a
contravariant interface as well). Alternative names could be:
- `IObservableSink1` (such that a write side could be named
`IObservableSource`).
- `IObservableReceiver` (such that a write side could be named
`IObservableSender`).

I'm open to any naming choices.