Viewer: Switch from ShadowOnlyMaterial to BackgroundMaterial (#16707)
`ShadowOnlyMaterial` does not currently have a native wgsl
implementation, which means it has to be converted to glsl at runtime,
which is not ideal. Switching to `BackgroundMaterial` solves this, plus
it removes the dependency on `@babylonjs/materials`.

I also switched the skybox material over to `BackgroundMaterial` since
it can do what we need in a lighter weight way (unlit + reflection
blur). In doing so, it revealed a bug in `BackgroundMaterial`'s wgls
fragment shader implementation, where it was simply missing `uniforms.`
when referencing the `vReflectionMicrosurfaceInfos` uniform.

This PR also includes a few other small Viewer changes related to
shadows:
- Store the shadow light as type `ShadowLight` instead of `Spotlight`.
This is really just in preparation to switch to `DirectionalLight`,
which will be a different PR.
- Set `ShadowGenerator.enableSoftTransparentShadow = true;` (feedback
from @Popov72).
- Set `ShadowGenerator.bias = radius / 1000;` (feedback from @Popov72,
and we'll see if we need to also allow the user to explicitly override
the `bias`).