Viewer: Add missing element event for shadows config (#16712)
This adds an event at the element level for the shadow config changing.
This just got missed in the initial shadows PR. However, I also noticed
that with some other refactors and merges, we lost the code that was
firing a couple other events, so added those back as well.