Inspector v2: Sandbox integration (#16717)
This is an initial integration of the inspector v2 into Sandbox. Some
notes on this:
- Using the `inspectorv2` query param (distinct from the `newux` param
@georginahalpern is using for the UI overhaul work, so that we can
enable/disable these separately).
- In a later PR (probably after a bit more inspector v2 feature work is
done), we'll add a button in the Sandbox toolbar that adds the query
param for you.
- Inspector v2 does not have debugLayer integration yet, so the code
uses the inspector v2 API directly.
- While the code change is small, the inspector v2 integration in
sandbox is quite different from v1 in terms of the bundle. For v1, the
inspector umd bundle is dynamically loaded (legacy approach). For v2,
inspector is an esm sandbox bundle chunk (modern approach).

![image](https://github.com/user-attachments/assets/3f6730da-7c1a-44c4-a217-2ec7309e2398)