FrameGraph: better support for autoCalcDepthBounds in cascaded shadow generator task (#16734)
Improved fix for #16710:
* The depth texture to use for min/max reduction is now an input of the
CSM task. This means that this depth texture can be shared by multiple
tasks. In the previous PR, a new depth texture was always generated
internally when `autoCalcDepthBounds` was **true**
* The textures used for min/max reduction are created as part of the
frame graph, which means that these textures can be optimized and reused
by the graph

I have also added support for “normalized view depth” as an output of
the geometry renderer task, as this is the most optimal type for min/max
depth reduction when used in CSM.