glTF Exporter: Fix light direction calculation (#16721)
The previous light direction calculation didn't account for when the
Babylon light direction matched glTF's. This would result NaN values in
RotationAxisToRef, as the rotation axis would be (0,0,0), and its length
(0) used as a divisor.

I noticed we already have a function to do what we need for the light
direction calculation-- `FromUnitVectorsToRef`-- so I've replaced the
whole thing with that. Much simpler.

(Bonus: I also noticed that the lights in our glTF light export test
don't use glTF falloff, so I've earmarked that change in this PR.)