Inspector v2: Test app (#16732)
This PR introduces a test app for local development of Inspector v2. It
uses webpack, but a newer version than our other tools which is needed
for compatibility with the webpack React fast refresh plugin (which is a
new local dev only dependency). The simple test app just loads a model,
and also enables drag & drop to easily test other models. Inspector is
always open, and fast refresh is supported so as you make changes to
React components, you immediately see them in the test app, without
reloading the page or re-creating the scene. You can run and debug the
test app from VSCode by just running the new "Inspector v2 development
(Chrome)" launch command. This enables a very tight dev loop. Hopefully
in the future we can see what it would take to bring fast refresh to
other tools, like Sandbox and Playground, but for now this is a low cost
solution to help speed up Inspector v2 development.