Inspector v2: Playground integration (#16741)
This is an initial integration of the inspector v2 into Playground. Some
notes on this:

- Using the inspectorv2 query param (distinct from the newux param
@georginahalpern is using for the UI overhaul work, so that we can
enable/disable these separately).
- In a later PR (probably after a bit more inspector v2 feature work is
done), we'll add a button in the Playground toolbar that adds the query
param for you.
- Inspector v2 does not have debugLayer integration yet, so the code
uses the inspector v2 API directly. Additionally, we'll want Playground
to be able to use new inspector v2 APIs (e.g. adding additional
Playground specific inspector features), in which case it will need to
directly call the inspector v2 APIs anyway.
- This required a change to prevent Fluent focus management from
stealing tab keys when focus is in the code editor (which should insert
tabs, not move focus!). More details in the comments.
- Inspector v2 does not have "embed" mode yet (scene explorer stacked on
top of the right pane), so in Playground it shows up as a left pane and
a right pane for now.

![image](https://github.com/user-attachments/assets/b831ce14-11bd-4521-9d8b-a049a4ce860b)