Add faster synchronous method for VAT baking on vertexAnimationBaker (#16749)
Rather than defer to the engine loop, we should be able to "stop the
world" and render each frame of a given range set to produce a VAT as
quickly as possible. Theoretically this could run on a worker with a
NullEngine and isolated scene, but figure the overhead might not be
worth spinning that up.

Add a test to match sandbox from
https://playground.babylonjs.com/?BabylonToolkit#CP2RN9#302

I don't know if @RaggarDK those assets are okay to check in and use in
the test suite/be available in the playground?

Added some perf logging in the tests with output on my machine

  console.log
    Synchronous bake took: 34.68687499999987 ms
  console.log
    Asynchronous bake took: 2059.416625 ms


It would be nice to see something like a "ThinContainerManager" that can
import .babylon,.gltf, etc. and do all the perf things under the hood
like create VAT, discard unused skeleton/AG, prime itself to create
instances... These benchmarks move towards being able to do something
like that in runtime.