Inspector v2: Factor react components out of service definitions (#16754)
Previously there were still some react components that were defined
"inline" in the service definitions. I'm factoring the bulk of these out
in this PR for a few reasons:
1. Had feedback from @georginahalpern and @sebavan that it would be
easier to understand/read/maintain if they were separated.
2. React fast refresh **only** works for components that are directly
exported from a module. If they are defined as part of an object, or
returned from a function, or anything other than a straight component
export, fast refresh does not work. So making this the standard pattern
will significantly improve the dev workflow.
3. It makes it much easier to re-use the bulk of the react components
part of things. For example, if we have a second hierarchy like view for
audio (separate from scene explorer, as @docEdub prototyped), or if
other tabs (like status and debug) have a similar UX to properties with
an accordion view, sections, etc.

As part of this, the raw react components got moved to the components
folder, and now what is in the services folder is the service
definitions that mostly are just managing the service registration and
service interactions.