Inspector v2: Stub out stats pane (#16755)
In this PR, I'm stubbing out a new stats pane, which includes:
1. Factoring out an `AccordionPane` react component from the
`PropertiesPane` component. The idea is that we will have several panes
that will have the same basic accordion UI, and each can support the
same type of extensibility as properties.
2. `PropertiesPane`/`PropertiesService` are slightly more complicated
that the other panes in that the "context" of the pane can have
different types (e.g. a Mesh or a Light). Given this, I kept the
predicate stuff specific to Properties. Other panes like Stats can have
a single type that is always the same (like Scene) and be a little
simpler.
3. Added a new `StatsPane`/`StatsService` that otherwise follow the same
model as properties.

One other difference that I am trying with `StatsService` is to not
bother having separate service definitions for the default/built-in
content of the status pane. Instead it is defined directly in the
`StatsService`, but new service definitions could be defined to extend
the default content of the stats pane. This was based off a suggestion
@deltakosh had in the original inspector v2 PR.


![image](https://github.com/user-attachments/assets/4c177ec7-e988-4eb5-8875-b423d93926c1)