Fixing issue when cloning LineMesh with CreationOptions (#16762)
When cloning a LineMesh, if the cloning is using MeshCloneOptions,
LineMesh will not clone correctly because it will not update the source
of the clone, and it will use the parent itself. This updates the
function to handle MeshCloneOptions the same way that mesh.ts->clone
does, where there are two code paths depending if parent is a Node or
MeshCloneOptions.