Remove dev step from tooling builds (reducing memory pressure) and copy the minified file to a max.js file for backwards compat (#16760)
Originally our build commands for our tools were running both dev and
prod, outputted an unminified max.js file during CI. This impacted
memory usage during build and is not necessary for debugging since we
offer source maps.

This PR removes the dev step from our builds, so we won't be outputting
the unminified max.js files anymore. But in order to preserve backwards
compatibility for users who may have been referencing the .max.js file
directly, we now copy the minified file to a .max.js file after the
build is complete. This plugin will only run if the `minToMax` option is
set to true in the webpack configuration.

---------

Co-authored-by: Georgina <gehalper@microsoft.com>