Inspector v2: Handle extension install errors (#16807)
A few people recently got hung up on extension install errors not
currently being handled, particularly in the case where a previously
installed extension can no longer be installed (such as the name
changing). This PR adds error handling for these scenarios that works
for our dev workflow as well as users.

Also unrelated, but I updated the theme with a color ramp generated from
the color used for icons in inspector v1. We can continue to tune of
course, I just wanted to put something in place that is slightly closer
to the Babylon colors.


![image](https://github.com/user-attachments/assets/e3d66afd-c4f0-472d-82d6-92cf5cb56caa)