Add audio `setVolume` method with ramp duration and curve shape options (#16781)
Adds the ability to set the volume of sounds and buses using a ramp
(a.k.a. a fade).

Note that due to a limitation in Firefox, ramps are not allowed to
overlap on the timeline, and this change causes and error to be thrown
if a volume ramp duration is set on a sound or bus that is already
ramping.

This change also moves common functionality into a new base class named
`AbstractAudioOutNode` and improves audio testing to handle error cases
and fixes flaky tests.