Inspector v2: Allow AccordionPane to have default content and use it in DebugService and SettingsService (#16813)
This change makes it so that `AccordionPane` can have default sections
with default content, which is merged with the "add on" sections and
content. The default sections can optionally specify an identity which
allows them to be extended (e.g. more content added to the default
sections). This is trying to make it really easy to define all the
default UI, but still have it be extensible. Later I will probably
change section identities to just be strings rather than symbols and
remove the `identity` prop (the `title` will be the identity), but I
don't want to make a breaking change right now since there are a bunch
of active PRs using it as is.