[Fluent] Scale slider min/max to avoid tick marks (#16819)
The Fluent slider uses a step size of 1 by default, but if the `step`
prop is set to anything other than `undefined`, it adds tick marks. This
is especially bad when you have something like min=0 max=1 step=0.01, as
it will render 100 tick marks, which basically makes the slider look
like garbage. This change just scales the min/max such that step can
always be undefined (the default value of 1) and never draw tick marks.

It also rounds the Input box numeric value to the closes step for
consistency.