Inspector v2: Allow target to be null or undefined in useProperty (and related functions) (#16824)
Previously we were not allowing for the case where the target object is
null or undefined. This happens in cases like the rotationQuaternion
property, which can be undefined. If it is, then the follow up hooks for
intercepting the x/y/z properties of the quaternion fail at runtime,
because the target was undefined. React hooks cannot be conditional, so
they must handle the null/undefined case.

This also required tightening up the typing, such that if target is
known at compile time to not be null, then the resulting property will
not be null (unless the property itself is allowed to be null). If
target might be null at compile time, then the resulting property value
is always possibly null. This makes usage easy (like with BoundProperty,
where the value might be expected to not be null or undefined).