Fix static sound disconnect edge case (#16846)
In rare cases when lots of static sounds are playing, it's possible for
static sound instances to have their `_onEnded` handler called by the
underlying AudioBufferSourceNode *after* the sound has been disposed.
This causes the static sound instance `_disconnect` call to occur twice,
which throws an error on the 2nd call since it is already disconnected.

This change adds an `_isConnected` flag to static sound instances which
skips the 2nd call to `_disconnect` if it is already disconnected.